+++ go/test/endtoend/vtgate/deletetest/delete_test.go
